refactor(hooks): share one dev-server classifier between the two dev hooks
pre-bash-dev-server-block.js refuses a dev server outside tmux and
auto-tmux-dev.js is what puts it in tmux, so a command one recognises and the
other does not is the worst outcome: refused, and never detached.

Both carried a copy of the same raw-text regex, each with a comment asking the
other to stay byte-identical — which is exactly how the previous commit left
them, since widening the blocker to quoted and option-separated spellings did
not reach the wrapper.

The classification now lives in scripts/lib/dev-command.js and both hooks call
it. pre-bash-dev-server-block.js drops from 304 lines to 34.
